+++ templates/invalid_guest.php
@@ -1,0 +1,6 @@
+<?php
+script('cookbook', 'guest');
+?>
+
+<div id="app">
+</div>